Prevent autofocus of textarea in Chrome.
Chrome automatically focuses a textarea once you retrieve it's cursor index. And as if that what's weird enough, in order to get Chrome to behave properly, I had to first `.blur()` the current textarea, and then `.focus()` the currently active element. Normally `.focus()` alone should suffice, but not here for some strange reason.